Fetch fee from envoy
